+++ Wiki/Azure/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [2021.12.08]
 
+## Updated
+
+- Updated `Create-Function-App.ps1` with added functionality for setting the default CORS settings for deployment slots.
+
 ## Breaking Changes
 
 - Changed the default route all traffic throught vnet behaviour for Application service with vnet integration(`Add-VNet-integration-to-Function-App.ps1`). The `RouteAllTrafficThroughVnet` is set to `false` now.